design/backend_practice_1/pyproject.toml
[tool.poetry]
name = "step1"
version = "0.1.0"
description = ""
authors = ["OPAYA <juntaegim99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.1"
dependency-injector = "^4.41.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"